Move to Go modules (#113)
* Improve logging for exec a bit


* Move to Go modules

Use Go modules and fix a few things when updating the go-github package.

+++ merge_handler.go
@@ -137,7 +137,7 @@ func (d Dreck) pullRequestDeletePendingReviews(client *github.Client, req types.
 
 	for _, review := range reviews {
 		// don't care about return code here.
-		if _, _, err := client.PullRequests.DeletePendingReview(ctx, req.Repository.Owner.Login, req.Repository.Name, int64(pull.GetNumber()), review.GetID()); err != nil {
+		if _, _, err := client.PullRequests.DeletePendingReview(ctx, req.Repository.Owner.Login, req.Repository.Name, pull.GetNumber(), review.GetID()); err != nil {
 			log.Warningf("Failed to delete pending review: %s", err)
 		}
 	}
